Add Progress Items in ProgressScreen

+++ app/src/main/java/com/example/flashcard/model/entities/SessionEntity.kt
@@ -20,5 +20,5 @@ data class SessionEntity(
 	var cardsReviewed: Long = 0,
 	@ColumnInfo(name = "cards_failed")
 	var cardsFailed: Long = 0,
-	var score: Long = 0,
+	var score: Int = 0,
 )
